requirements.txt
alembic==1.10.4
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.0.8
argh==0.28.1
Authlib==1.2.0
Babel==2.12.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.2
debtcollector==2.5.0
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.4.0
docker==6.0.1
dogpile.cache==1.2.2
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Admin==1.6.1
flask-babel==3.1.0
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-ReCaptcha==0.4.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
humanize==4.6.0
idna==3.4
importlib-resources==6.0.0
iso8601==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
keystoneauth1==5.2.1
libvirt-python==9.3.0
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
msgpack==1.0.5
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==1.1.0
ordered-set==4.1.0
os-service-types==1.7.0
oslo.i18n==6.0.0
oslo.serialization==5.1.1
oslo.utils==6.2.0
packaging==23.1
paramiko==3.2.0
pbr==5.11.1
prettytable==3.8.0
psycopg2==2.9.6
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.15.1
PyNaCl==1.5.0
pyparsing==3.1.0
python-cinderclient==9.3.0
python-novaclient==18.3.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requestsexceptions==1.4.0
rich==13.4.2
sentry-sdk==1.24.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
stevedore==5.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
waitress==2.1.2
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.3.3
wrapt==1.15.0
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
